unified-ml-endpoint/pyproject.toml
# Poetry
[tool.poetry]
name = "sai-unified-ml-endpoint"
version = "0.0.1"
description = "This is Tech CME GenAI PoC repo"
authors = ["tbd"]

[tool.poetry.dependencies]
python = "~3.12"
nbstripout = "0.7.1"
langchain = "0.3.21"
langchain-core = "0.3.49"
openai = "^1.81.0"
transformers = "4.44.2"
wget = "3.2"
pypdf = "4.3.1"
PyPDF2 = "3.0.1"
faiss-cpu = "1.8.0.post1"
retrying = "1.3.4"
nltk = "3.9.1"
sentence-transformers = "3.0.1"
ffmpeg-python = "0.2.0"
tenacity = "8.5.0"
pymilvus = "2.4.6"
tiktoken = "0.7.0"
numba = "0.60.0"
openpyxl = "3.1.5"
python-pptx = "1.0.2"
python-docx = "1.1.2"
python-box = "7.2.0"
pip-shims = "0.7.3"
st-pages = "1.0.1"
Pillow = "10.4.0"
pymupdf = "1.24.10"
tabula-py = "2.9.3"
tabula = "1.0.5"
pyyaml = "6.0.2"
cryptography = "43.0.1"
pyopenssl = "24.2.1"
psycopg2-binary = "2.9.9"
pgvector = "0.3.2"
sqlalchemy = "2.0.34"
mlflow = "2.16.0"
scikit-learn = "1.5.1"
spacy = "3.7.6"
fuzzywuzzy = "0.18.0"
numpy = "1.26"
jinja2 = "3.1.4"
numerize = "0.12"
langchain-openai = "0.2.3"
langchain-community = "0.3.3"
azureml-core = "1.57.0.post1"
requests = "2.32.3"
langgraph = "0.2.39"
python-levenshtein = "0.25.1"
alembic = "1.13.3"
httpx = "0.27.2"
langchain-experimental = "0.3.3"
marshmallow = "3.23.1"
snowflake-connector-python = "3.15.0"
snowflake-sqlalchemy = "1.7.3"
uvicorn = "^0.34.0"
fastapi = "^0.115.8"
boto3 = "^1.37.3"
duckduckgo-search = "7.5.4"
bs4 = "^0.0.2"
rapidfuzz = "3.12.1"
loguru = "^0.7.3"
python-jose = "^3.5.0"

[tool.poetry.dev-dependencies]
isort = "5.12.0"
pydocstyle = "6.3.0"
pylint = "2.16.3"
black = "23.1.0"
yamllint = "1.29.0"
pytest = "8.1.1"     # updated library on march 2024
pytest-cov = "5.0.0" # updated library on march 2024

en_core_web_sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl" }
en_core_web_lg = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl" }

# Formatting
[tool.isort]
profile = "black"

[tool.pytest.ini_options]
pythonpath = ["src"]
markers = ["e2e_azure", "e2e_local", "components", "utils"]
filterwarnings = ["ignore::DeprecationWarning::", "ignore::FutureWarning::"]

[tool.pydocstyle]
convention = "google"

# Testing
[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

# Omit packages in test coverage
[tool.coverage.run]
omit = ["*/azml_components/*", "*/azml_realtime_deployments/*", "*/utils/*"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 130
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''